Ensure terminating null is included when setting values via RegSetValueEx
Although values set using RegSetValueEx without inclusion of the terminating null in the
"cbData" parameter seem to still get null terminated by the system (based upon inspection
of the binary data in Registry Editor), the MSDN page for RegSetValueEx is quite clear
that the terminating null should be included in "cbData". Indeed, it is already included
in one place in EnvVar.cpp. This change ensures that all calls to RegSetValueEx include
the terminating null in "cbData".

h/t Debra Do